# Conflicts:
#	src/main/java/lapr/project/gpsd/controller/CreateServiceRequestController.java
#	src/main/java/lapr/project/gpsd/model/ServiceRegistry.java

+++ src/main/java/lapr/project/gpsd/model/ServiceRequestDescription.java
@@ -9,7 +9,7 @@ public class ServiceRequestDescription {
      */
     private Service service;
     private String description;
-    private double duration;
+    private int duration;
     private boolean assigned;
 
     /**
@@ -18,7 +18,7 @@ public class ServiceRequestDescription {
      * @param description
      * @param duration 
      */
-    public ServiceRequestDescription(Service service, String description, double duration) {
+    public ServiceRequestDescription(Service service, String description, int duration) {
         this.service = service;
         this.description = description;
         this.duration = duration;
